Refactor source
