+++ .gitignore
@@ -13,7 +13,6 @@
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
-
 # Dependency directories (remove the comment below to include it)
 # vendor/
 
@@ -22,3 +21,6 @@ go.work
 .idea/
 data/
 local/data/
+
+run.sh
+scratch/
