+++ Dockerfile
@@ -1,15 +1,15 @@
 FROM golang:alpine3.19 as build
 
-WORKDIR /certificate.stream
+WORKDIR /build
 COPY . .
 
 RUN apk update && apk add git \
     && go mod tidy \
-    && go build -o certificate .
+    && go build -o ctlog .
 
 FROM alpine:3.19
 
-ENV VERSION=0.1.1
+ENV VERSION=0.1.2
 
 ARG CREATED
 ARG REVISION
@@ -19,16 +19,16 @@ LABEL org.opencontainers.image.description="Certificate Transparency Log monitor
 LABEL org.opencontainers.image.version=$VERSION
 LABEL org.opencontainers.image.authors="pogzyb@umich.edu"
 LABEL org.opencontainers.image.url="https://github.com/pogzyb/certificate.stream"
-LABEL org.opencontainers.image.source="https://github.com/pogzyb/certificate.stream/certificate.stream"
+LABEL org.opencontainers.image.source="https://github.com/pogzyb/certificate.stream"
 LABEL org.opencontainers.image.documentation="https://github.com/pogzyb/certificate.stream"
 LABEL org.opencontainers.image.created=$CREATED
 LABEL org.opencontainers.image.revision=$REVISION
 LABEL org.opencontainers.image.licenses="MIT"
 
-COPY --from=build /certificate.stream/certificate /usr/local/bin/certificate
-RUN chmod u+x /usr/local/bin/certificate
+COPY --from=build /build/ctlog /usr/local/bin/ctlog
+RUN chmod u+x /usr/local/bin/ctlog
 
 USER guest
 
-ENTRYPOINT [ "certificate" ]
+ENTRYPOINT [ "ctlog" ]
 CMD [ "--help" ]
