+++ sink/firehose/put.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"os"
 
-	"certificate.stream/service/pkg/certificate/v1"
+	"certificate.stream/service/certificate/v1"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/config"
