+++ stream/loglist.go
